requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==20.3.0
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
check50==3.3.11
click==8.1.7
colorama==0.4.6
cryptography==42.0.3
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2024.2.0
gast==0.4.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
jellyfish==0.11.2
Jinja2==3.1.3
joblib==1.3.2
keras==2.13.1
lib50==3.0.11
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pexpect==4.9.0
pillow==10.2.0
protobuf==4.25.3
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.5.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soupsieve==2.5
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
threadpoolctl==3.3.0
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.38.1
typing_extensions==4.5.0
urllib3==2.2.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
